% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmfit2table.R
\name{glmfit2table}
\alias{glmfit2table}
\title{glmfit2table}
\usage{
glmfit2table(a, covariate)
}
\arguments{
\item{a}{A glm.fit() object}

\item{covariate}{A covariate name}
}
\value{
Returns a given coefficent's beta and p value
}
\description{
Return beta and p value for selected coefficent from glm.fit()
}

