% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heart_disease}
\alias{heart_disease}
\title{Heart disease}
\format{A data frame with 302 rows and 14 variables:
\describe{
  \item{age}{age of patient (29--77)}
  \item{sex}{sex of patient (0 (#####),1 (#####))}
  \item{cp}{####### (from 1 (worst####) to 4 (best####))}
  \item{trestbps}{###### (94--200)}
  \item{chol}{##### (126--564)}
  \item{fbs}{##### (0,1)}
  \item{restecg}{###### (0,2)}
  \item{thalach}{#### (71-202)}
  \item{exang}{##### (0,1)}
  \item{oldpeak}{####### (0--6.2)}
  \item{slope}{#### (1,2,3)}
  \item{ca}{###### (1 to 5)}
  \item{thal}{#### (1 to 4)}
  \item{num}{#### (0,1)}
}}
\usage{
heart_disease
}
\description{
A dataset containing #### data for 302 patients.
The variables are as follows:
}
\keyword{datasets}
