% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearfragility.R
\name{linearfragility}
\alias{linearfragility}
\title{Linear Fragility Function}
\usage{
linearfragility(formula, data, covariate = "all.factors.default",
  conf.level = 0.95, verbose = FALSE)
}
\arguments{
\item{formula}{Model formula which will be evaluated by lm()}

\item{data}{Dataframe with values for model forma, passed to lm()}

\item{covariate}{Vector of covariates to find fragility index for. Default is all covariates in formula}

\item{conf.level}{Significance level}

\item{verbose}{Logical indicating if function will return verbose results or only fragility index}
}
\value{
If verbose is FALSE, returns a list with fragility indices for selected covariates. If
verbose is TRUE, returns a list with p-values for each fragility index at each iteration
of the algorithm.
}
\description{
Compute the fragility of a coefficient in a linear regression, i.e. the number of removed observations it would take to make a significant-result non-significant. Uses the lm() function from the stats package.
}
