% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_genes}
\alias{tcga_genes}
\title{CNA patient survival}
\format{A data frame with 276 rows and 3 variables:
\describe{
  \item{months}{months ##### (0.92--72.01)}
  \item{status}{######## (0,1)}
  \item{strat}{####### (mutated, not_mutated)}
}}
\usage{
tcga_genes
}
\description{
A dataset containing ######## data for 276 patients.
The variables are as follows:
}
\keyword{datasets}
